backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1

# RAG Stack
langchain==0.1.9
langchain-community==0.0.21
langchain-openai==0.0.5
chromadb==0.4.22
sentence-transformers==2.3.1

# Anthropic Claude API
anthropic==0.18.0
requests==2.31.0

# Database
sqlalchemy==2.0.25

# Code Processing
gitpython==3.1.41
tree-sitter==0.21.0
tree-sitter-python==0.23.6
tree-sitter-javascript==0.23.1

# Utilities
aiofiles==23.2.1
python-multipart==0.0.9

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0